OGC Refactoring Part 2 - Combine OgcDistributionsApi with OgcDistributionsApiController
