+++ server/src/main/java/de/ingrid/igeserver/api/OgcDistributionsApiController.kt.kt
@@ -20,12 +20,16 @@
 package de.ingrid.igeserver.api
 
 import de.ingrid.igeserver.services.OgcDistributionsService
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.Parameter
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.responses.ApiResponses
 import org.apache.logging.log4j.kotlin.logger
 import org.springframework.context.annotation.Profile
+import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.Authentication
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import org.springframework.web.multipart.MultipartFile
 
 @RestController
@@ -33,28 +37,40 @@ import org.springframework.web.multipart.MultipartFile
 @RequestMapping(path = ["/api/ogc"])
 class OgcDistributionsApiController(
     private val ogcDistributionsService: OgcDistributionsService,
-): OgcDistributionsApi {
+) {
 
     val log = logger()
 
-    override fun postDistribution(
-        allHeaders: Map<String, String>,
+    @PostMapping("/collections/{collectionId}/items/{recordId}/distributions", consumes = [MediaType.MULTIPART_FORM_DATA_VALUE]) //"multipart/form-data",
+    @Operation(tags=["OGC/Distributions"], responses = [], summary = "Upload distributions (files) to a specific collection and record", hidden = false )
+    @ApiResponses(value = [
+        ApiResponse(responseCode = "200", description = "Distributions uploaded successfully"),
+        ApiResponse(responseCode = "400", description = "Invalid request")
+    ])
+    fun postDistribution(
+        @RequestHeader allHeaders: Map<String, String>,
         principal: Authentication,
-        collectionId: String,
-        recordId: String,
-        files: List<MultipartFile>
+        @Parameter(description = "## Collection ID \n **OGC Parameter** \n\n The identifier for a specific record collection (i.e. catalogue identifier).", required = true) @PathVariable("collectionId") collectionId: String,
+        @Parameter(description = "## Record ID \n **OGC Parameter** \n\n The identifier for a specific record (i.e. record identifier).", required = true) @PathVariable("recordId") recordId: String,
+        @Parameter(description = "File the should be uploaded.", required = true) @RequestPart files: List<MultipartFile>,
     ): ResponseEntity<String> {
         val userID = principal.name
         ogcDistributionsService.handleUploadDistribution(principal, userID, collectionId, recordId, files)
         return ResponseEntity.ok().build()
     }
 
-    override fun deleteDistribution(
-        allHeaders: Map<String, String>,
+    @DeleteMapping("/collections/{collectionId}/items/{recordId}/distributions")
+    @Operation(tags=["OGC/Distributions"], responses = [], summary = "Delete an existing distribution (file)", hidden = false )
+    @ApiResponses(value = [
+        ApiResponse(responseCode = "200", description = "Distribution deleted successfully"),
+        ApiResponse(responseCode = "400", description = "Invalid request")
+    ])
+    fun deleteDistribution(
+        @RequestHeader allHeaders: Map<String, String>,
         principal: Authentication,
-        collectionId: String,
-        recordId: String,
-        distributionId: String
+        @Parameter(description = "## Collection ID \n **OGC Parameter** \n\n The identifier for a specific record collection (i.e. catalogue identifier).", required = true) @PathVariable("collectionId") collectionId: String,
+        @Parameter(description = "## Record ID \n **OGC Parameter** \n\n The identifier for a specific record (i.e. record identifier).", required = true) @PathVariable("recordId") recordId: String,
+        @Parameter(description = "## distribution ID \n\n The Identifier of the distribution.") @RequestParam(value = "uri", required = true) distributionId: String,
     ): ResponseEntity<String> {
         val userID = principal.name
         ogcDistributionsService.handleDeleteDistribution(principal, userID, collectionId, recordId, distributionId)
